+++ visualization/app/codeCharta/ui/ribbonBar/heightSettingsPanel/heightSettingsPanel.module.ts
@@ -1,0 +1,14 @@
+import { CommonModule } from "@angular/common"
+import { NgModule } from "@angular/core"
+import { MaterialModule } from "../../../../material/material.module"
+import { ResetSettingsButtonModule } from "../../resetSettingsButton/resetSettingsButton.module"
+import { SliderModule } from "../../slider/slider.module"
+import { HeightSettingsPanelComponent } from "./heightSettingsPanel.component"
+
+@NgModule({
+	imports: [CommonModule, MaterialModule, SliderModule, ResetSettingsButtonModule],
+	declarations: [HeightSettingsPanelComponent],
+	exports: [HeightSettingsPanelComponent],
+	entryComponents: [HeightSettingsPanelComponent]
+})
+export class HeightSettingsPanelModule {}