+++ visualization/app/codeCharta/ui/ribbonBar/heightSettingsPanel/selectors/isLabelsSliderDisabled.selector.ts
@@ -1,0 +1,7 @@
+import { createSelector } from "../../../../state/angular-redux/createSelector"
+import { colorLabelsSelector } from "../../../../state/store/appSettings/colorLabels/colorLabels.selector"
+
+export const isLabelsSliderDisabledSelector = createSelector(
+	[colorLabelsSelector],
+	colorLabels => colorLabels.negative || colorLabels.neutral || colorLabels.positive
+)