+++ visualization/app/codeCharta/ui/slider/slider.component.ts
@@ -1,6 +1,7 @@
 import "./slider.component.scss"
 import { Component, Input } from "@angular/core"
 import { MatSliderChange } from "@angular/material/slider"
+import { parseNumberInput } from "./util/parseNumberInput"
 
 @Component({
 	selector: "cc-slider",
@@ -10,7 +11,9 @@ export class SliderComponent {
 	@Input() value?: number
 	@Input() min: number
 	@Input() max: number
+	@Input() label: string
 	@Input() step?: number = 1
+	@Input() disabled?: boolean = false
 	@Input() onChange: (number) => void
 
 	handleSliderOnChange($event: MatSliderChange) {
@@ -20,8 +23,8 @@ export class SliderComponent {
 	}
 
 	handleInputOnChange($event: InputEvent) {
-		const newValue = Number.parseInt(($event.target as HTMLInputElement).value)
-		if (newValue !== this.value) {
+		const newValue = parseNumberInput($event, this.min, this.max)
+		if (newValue !== this.value && !Number.isNaN(newValue)) {
 			this.onChange(newValue)
 		}
 	}