+++ visualization/app/codeCharta/util/aggregationGenerator.ts
@@ -19,7 +19,7 @@ export class AggregationGenerator {
 
 		for (const inputFile of inputFiles) {
 			this.projectNameArray.push(inputFile.fileMeta.projectName.replace(" ", "_"))
-			this.fileNameArray.push(FileNameHelper.withoutCCJsonExtension(inputFile.fileMeta.fileName).replace(" ", "_"))
+			this.fileNameArray.push(FileNameHelper.withoutCCExtension(inputFile.fileMeta.fileName).replace(" ", "_"))
 			this.fileChecksumArray.push(inputFile.fileMeta.fileChecksum)
 			this.fileSizesSum += inputFile.fileMeta.exportedFileSize
 		}