+++ visualization/app/codeCharta/util/customConfigHelper.ts
@@ -1,5 +1,5 @@
 "use strict"
-import { ColorRange, LocalStorageCustomConfigs, stateObjectReplacer, stateObjectReviver } from "../codeCharta.model"
+import { LocalStorageCustomConfigs, stateObjectReplacer, stateObjectReviver } from "../codeCharta.model"
 import { CustomConfigItemGroup } from "../ui/customConfigs/customConfigs.component"
 import {
 	CustomConfig,
@@ -14,12 +14,11 @@ import { setState } from "../state/store/state.actions"
 import { setColorRange } from "../state/store/dynamicSettings/colorRange/colorRange.actions"
 import { setMargin } from "../state/store/dynamicSettings/margin/margin.actions"
 import { setCamera } from "../state/store/appSettings/camera/camera.actions"
-import { Vector3 } from "three"
 import { setCameraTarget } from "../state/store/appSettings/cameraTarget/cameraTarget.actions"
-import { StoreService } from "../state/store.service"
 import { ThreeCameraService } from "../ui/codeMap/threeViewer/threeCameraService"
 import { ThreeOrbitControlsService } from "../ui/codeMap/threeViewer/threeOrbitControlsService"
 import { BehaviorSubject } from "rxjs"
+import { Store } from "../state/angular-redux/store"
 
 export const CUSTOM_CONFIG_FILE_EXTENSION = ".cc.config.json"
 const CUSTOM_CONFIGS_LOCAL_STORAGE_VERSION = "1.0.1"
@@ -29,46 +28,6 @@ export const CUSTOM_CONFIGS_LOCAL_STORAGE_ELEMENT = "CodeCharta::customConfigs"
 export class CustomConfigHelper {
 	private static customConfigs: Map<string, CustomConfig> = CustomConfigHelper.loadCustomConfigs()
 	static customConfigChange$: BehaviorSubject<null> = new BehaviorSubject(null)
-
-	static getCustomConfigItemGroups(customConfigFileStateConnector: CustomConfigFileStateConnector): Map<string, CustomConfigItemGroup> {
-		const customConfigItemGroups: Map<string, CustomConfigItemGroup> = new Map()
-
-		for (const customConfig of CustomConfigHelper.customConfigs.values()) {
-			const groupKey = `${customConfig.assignedMaps.join("_")}_${customConfig.mapSelectionMode}`
-
-			if (!customConfigItemGroups.has(groupKey)) {
-				customConfigItemGroups.set(groupKey, {
-					mapNames: customConfig.assignedMaps.join(" "),
-					mapSelectionMode: customConfig.mapSelectionMode,
-					hasApplicableItems: false,
-					customConfigItems: []
-				})
-			}
-
-			const customConfigItemApplicable = CustomConfigHelper.isCustomConfigApplicable(customConfigFileStateConnector, customConfig)
-			customConfigItemGroups.get(groupKey).customConfigItems.push({
-				id: customConfig.id,
-				name: customConfig.name,
-				mapNames: customConfig.assignedMaps.join(" "),
-				mapSelectionMode: customConfig.mapSelectionMode,
-				isApplicable: customConfigItemApplicable
-			})
-
-			if (customConfigItemApplicable) {
-				customConfigItemGroups.get(groupKey).hasApplicableItems = true
-			}
-		}
-
-		return customConfigItemGroups
-	}
-
-	private static isCustomConfigApplicable(customConfigFileStateConnector: CustomConfigFileStateConnector, customConfig: CustomConfig) {
-		// Configs are applicable if their mapChecksums (and mode) are matching, therefore, map names must not be checked.
-		return (
-			customConfigFileStateConnector.getChecksumOfAssignedMaps() === customConfig.mapChecksum &&
-			customConfigFileStateConnector.getMapSelectionMode() === customConfig.mapSelectionMode
-		)
-	}
 
 	static setCustomConfigsToLocalStorage() {
 		const newLocalStorageElement: LocalStorageCustomConfigs = {
@@ -80,7 +39,7 @@ export class CustomConfigHelper {
 		CustomConfigHelper.customConfigChange$.next(null)
 	}
 
-	private static loadCustomConfigs() {
+	static loadCustomConfigs() {
 		const ccLocalStorage = this.getCcLocalStorage()
 		return new Map(ccLocalStorage?.customConfigs)
 	}
@@ -221,7 +180,7 @@ export class CustomConfigHelper {
 			customConfigFileStateConnector.isEachFileSelected()
 		) {
 			// If only one map is uploaded/present in SINGLE mode, prefix the .cc.config.json file with its name.
-			fileName = `${FileNameHelper.withoutCCJsonExtension(customConfigFileStateConnector.getJointMapName())}_${fileName}`
+			fileName = `${FileNameHelper.withoutCCExtension(customConfigFileStateConnector.getJointMapName())}_${fileName}`
 		}
 
 		FileDownloader.downloadData(JSON.stringify(customConfigsDownloadFile, stateObjectReplacer), fileName)
@@ -287,7 +246,7 @@ export class CustomConfigHelper {
 
 	static applyCustomConfig(
 		configId: string,
-		storeService: StoreService,
+		store: Store,
 		threeCameraService: ThreeCameraService,
 		threeOrbitControlsService: ThreeOrbitControlsService
 	) {
@@ -298,13 +257,13 @@ export class CustomConfigHelper {
 
 		// TODO: Check if state properties differ
 		// Create new partial State (updates) for changed values only
-		storeService.dispatch(setState(customConfig.stateSettings))
+		store.dispatch(setState(customConfig.stateSettings))
 
 		// Should we fire another event "ResettingStateFinishedEvent"
 		// We could add a listener then to reset the camera
 
-		storeService.dispatch(setColorRange(customConfig.stateSettings.dynamicSettings.colorRange as ColorRange))
-		storeService.dispatch(setMargin(customConfig.stateSettings.dynamicSettings.margin))
+		store.dispatch(setColorRange(customConfig.stateSettings.dynamicSettings.colorRange))
+		store.dispatch(setMargin(customConfig.stateSettings.dynamicSettings.margin))
 
 		// TODO: remove this dirty timeout and set camera settings properly
 		// This timeout is a chance that CustomConfigs for a small map can be restored and applied completely (even the camera positions)
@@ -312,8 +271,8 @@ export class CustomConfigHelper {
 			threeCameraService.setPosition()
 			threeOrbitControlsService.setControlTarget()
 
-			storeService.dispatch(setCamera(customConfig.stateSettings.appSettings.camera as Vector3))
-			storeService.dispatch(setCameraTarget(customConfig.stateSettings.appSettings.cameraTarget as Vector3))
+			store.dispatch(setCamera(customConfig.stateSettings.appSettings.camera))
+			store.dispatch(setCameraTarget(customConfig.stateSettings.appSettings.cameraTarget))
 		}, 100)
 	}
 }