+++ CHANGELOG.md
@@ -12,13 +12,20 @@ and this project adheres to [Semantic Versioning](http://semver.org/)
 -   Add interactive dialog support for the parsers [#2737](https://github.com/MaibornWolff/codecharta/pull/2737) <br>
     ![m](https://user-images.githubusercontent.com/48621967/161549546-1463914e-c223-4912-acb1-db4e357e76c5.png)
 
+### Changed
+
+-   Remove all mentions of regular `.cc` `.json` `.gz` file extensions from file panel and sidebar [#2793](https://github.com/MaibornWolff/codecharta/pull/2793)
+
 ### Fixed 🐞
 
 -   Update slider of area metric options correctly on changes of related input field [#2787](https://github.com/MaibornWolff/codecharta/pull/2787)
+-   Prevent invalid input for margin in area metric options to be submitted [#2799](https://github.com/MaibornWolff/codecharta/pull/2799)
 
 ### Chore 👨‍💻 👩‍💻
 
 -   Introduce custom angular material theme [#2784](https://github.com/MaibornWolff/codecharta/pull/2784)
+-   Migrate height-settings-panel-component [#2790](https://github.com/MaibornWolff/codecharta/pull/2790)
+-   Fully migrate custom-config-component [#2760](https://github.com/MaibornWolff/codecharta/pull/2760)
 
 ## [1.95.2] - 2022-04-19